Rename the data type that is used to return the result of creating an instant notification order to the client
